Only define repository dependencies if they haven't been already.
This will improve the situation if multiple repositories define the
same dependencies (such as someone depending on rules_apple and
rules_swift, and then rules_apple also itself depending on
rules_swift). It does not address the situation where the same
dependency is used but with different versions; Bazel does not have
a great answer to this at this moment, as far as I know.