Use WORKSPACE relative file paths when using direct sources and WKTs
When using descriptor_set_in, protoc expects files as named when
creating it. While File.short_path is supposed to do the trick, it is
not working for extenal WORKSPACES. In turn, Well Known Types and
direct source build fail to compile because they are not recognized
by protoc.

Fixes #23

Signed-off-by: Steeve Morin <steeve.morin@gmail.com>