Refactor how the rules register actions that invoke the toolchain.
This is a major refactor of how actions are registered, by moving the action registration logic into partials that can capture toolchain-specific information. Now, instead of it being the individual callers' responsibilities to specify the correct wrapper, environment, and execution requirements, they simply ask the toolchain to register the action and it applies any requirements captured by the partial.

This greatly simplifies the logic for actions and makes it easier for custom rule authors who interact with the toolchain to spawn their own actions that use the toolchain's binaries, and it eliminates the distinction between when the toolchain *target* needs to be passed and when the toolchain *provider* is passed.

PiperOrigin-RevId: 202655061