`swift_test` emits .xctest bundles by default for toolchains that support Objective-C interop.
Fixes #32.

RELNOTES:
The behavior of `swift_test` now differs slightly for macOS targets, in order to provide
seamless integration with Apple's XCTest framework. The output of the rule is still a
binary, but one whose Mach-O type is `MH_BUNDLE` (a loadable bundle). Thus, the binary
cannot be launched directly. Instead, running `bazel test` on the target will launch a test
runner script that copies it into an `.xctest` bundle directory and then launches the
`xctest` helper tool from Xcode, which uses Objective-C runtime reflection to locate the
tests.

Test bundling on macOS can be disabled on a per-target basis, if desired. You may wish to
do this if you are not using XCTest, but rather a different test framework (or no framework
at all) where the pass/fail outcome is represented as a zero/non-zero exit code (as is the
case with other Bazel test rules like `cc_test`). To do so, disable the
`"swift.bundled_xctests"` feature on the target. You can also disable this feature for all
the tests in a package by applying it to your BUILD file's `package()` declaration.

PiperOrigin-RevId: 209952360