Add wrapper scripts for swift/swiftc.
Currently, only wrapper_swiftc is used; it provides an option to use an ephemeral
module cache directory in an attempt to get rid of flakiness with the default
module cache that causes compiler crashes after some configuration changes or
repository syncs.

RELNOTES: Use an ephemeral module cache when invoking `swiftc`.
PiperOrigin-RevId: 213985987