Add ability to use the Clang module cache from objc_library as a feature.
Moving the Clang module cache to an ephemeral directory caused noticeable
regressions in local build times. We add this capability back so that it
can be optionally adopted by passing the following flag to Bazel:

    --features=swift.use_global_module_cache

It is off by default since module cache corruption is known to cause
crashes for some users, so off-by-default is still the safer option. But
if you do not experience those crashes, then enabling this flag in your
global config can improve build performance.

PiperOrigin-RevId: 214677843