Remove `cc_libs` and `module_link_name` attributes from `swift_library`.
RELNOTES: The `cc_libs` and `module_link_name` attributes have been removed from the Swift rules. If you need to create a static library that contains both Swift and C++ code, have the `swift_library` depend on the `cc_library` and then use another rule, such as a `genrule`, to combine the archives.

To replace the behavior of the `module_link_name` attribute, pass the `-module-link-name <name>` flags directly in the target's `copts`. This change no longer affects the name of the output library; it is the responsibility of the user to ensure that the `.a` file that is output by the rule matches the link name embedded in the module. This allows the link name to refer to an archive emitted by a different rule, such as one that merges static archives as described above.

PiperOrigin-RevId: 233807274