Ensure that `SwiftInfo` providers aren't lost when passing through non-Swift targets.
For example, imagine a `swift_library(A)` depends on an `objc_library(B)` which depends on a `swift_library(C)`. Since `objc_library` doesn't know about `SwiftInfo`, it doesn't propagate it, and the modules from C don't get passed to the compilation action for A. This can sometimes cause odd build failures when the Swift compiler can't find modules that define types being used through those dependencies.

This change introduces an aspect that attaches to any dependencies of a `swift_library` target and ensures that all of them propagate a `SwiftInfo` (by merging the `SwiftInfo` providers of their dependencies, recursively).

RELNOTES: None.
PiperOrigin-RevId: 233980475