Travis CI Cleanup.
Bazel's BuildKite seems to be fine for the bazel builds, so remove the support
for doing bazel builds on Travis.

Note: Keeping the matrix as a manual build and some of the conditions in the
scripts just to reduce the chances of something not being right, as eventually
the BuildKite setup will match the buildifier support and these files will
just get dropped.

RELNOTES: None
PiperOrigin-RevId: 236655099