Don't eval tools for truth, just fail if it wasn't the right type.
Newer buildifiers were still flagging tools as not always being set,
but the real issue was if the value in the args was say an empty
string, then that last if would have still failed because an empty
string is False. Instead fail if someone passes anything that isn't
a list or depset no mater what the value is.

RELNOTES: None.
PiperOrigin-RevId: 241521323