Create a "dummy worker" for upstream integration tests.
The external dependencies used by the persistent worker are fairly heavyweight, and upstream integration tests (like those in rules_apple) should not be forced to mock them all out in their test workspace (which they would have to, because the worker is an implicit dependency of all toolchains). This change provides a flag that they can pass to Bazel to build a dummy executable that is dependency-free instead.

RELNOTES: None.
PiperOrigin-RevId: 242571122