Clean up clang usage for linking executables.
With all the new C++ APIs in Bazel, we should no longer be grabbing the clang executable separately from a legacy provider field. Instead, use cc_common to get the tool path and command line as we now do for the archiver.

RELNOTES: The `SwiftToolchainInfo.clang_executable` field has been removed. Users who were relying on this should now use the `cc_common` APIs to retrieve the tool for the desired CROSSTOOL action.
PiperOrigin-RevId: 250381190