Fix the `TOOLCHAINS` env var for custom Xcode toolchains.
This was previously set in a separate dictionary to prevent conflicts with static library archiving actions, but now those go through the C++ APIs and those conflicts no longer occur. Unfortunately, that change also broke this, because the new dictionary was no longer being used anywhere, so this variable never makes it through. The fix is to just set it on the original `env` dictionary again.

RELNOTES: None.
PiperOrigin-RevId: 274019010