Refactor slightly how actions are registered in the Swift rules.
Swift toolchains now have a `swift_executable` attribute that lets the configurator replace the Swift driver used by the rules with a custom executable in their workspace. This can be useful for testing the build rules alongside compiler development, and since that file is part of the workspace and a proper input of actions, it also works correctly across remote build environments.

Also remove the no-longer-used `SwiftToolchainInfo.clang_executable` field, now that linking has migrated to the `cc_common.link` API.

RELNOTES: None.
PiperOrigin-RevId: 281620086