Add `swiftc -version` as input to Linux actions
This fixes an issue where if you update the version of Swift on a Linux
machine you share caches with previous Swift versions which causes
compiler failures. Now the output of `swiftc -version` is included as an
input of all compile actions. This is analogous to how the Xcode version
is provided for macOS actions.