Migrate a number of feature-driven compilation flags to the action config model.
This change removes `swiftc_command_line_and_inputs` as a public API (there are no users currently within Google, which gives us an opportunity to rethink its design under the new model). Then, we move the majority of its feature checks that control command line flags to equivalent COMPILE action configs.

This got slightly awkward around the current treatment of WMO, since we need to check both the feature configuration and the user compile flags to determine what kind of outputs are needed, so in some cases, the check had to be split?for example, the action config already filters based on feature, so the configurator function needs to check the user compile flags for the presence of WMO based only if features for WMO *weren't* satisfied.

RELNOTES: None.
PiperOrigin-RevId: 289516791