Remove the `swift.index_while_building` feature.
This was never used; for each of the use cases we originally planned, we ended up going in different directions, so now it just adds complexity to the build rules. For local builds, just pass `-index-store-path` as a manual `--swiftcopt` to get an indexstore that is shared across all actions. For remote indexing, use an aspect that invokes the compiler in parallel to perform indexing (hand-waving details).

RELNOTES: None.
PiperOrigin-RevId: 289912580