Allow `.swiftinterface` files to be emitted without library evolution enabled by adding `--features=swift.emit_swiftinterface` to the build.
Note that doing so causes the compiler to emit a warning that `.swiftinterface` files are not supported without library evolution. However, the file is still emitted and has what appears to be somewhat valid content.

These interface files should not be used as inputs for upstream builds, and certainly not used for the creation of distributed libraries. The purpose of this change is to allow the creation of other source tooling that loads `.swiftinterface` files, when that tooling may be built from sources of the compiler that don't match the version that was used to build the module itself.

RELNOTES: None.
PiperOrigin-RevId: 295795663