Require the `swift_module` tag on `cc_library` targets that are (direct or transitive) dependencies of `swift_*` targets to generate module maps for them and import them into Swift.
This wasn't a hard requirement before because, even if we created a
module map for a library that didn't contain Swift-compatible C code,
as long as no Swift code actually tried to import it, no harm was
done. This will no longer be true when we start precompiling explicit
modules for C dependencies, at which point we **must** be able to
distinguish which ones are Swift-compatible and which are not (i.e.,
currently most of them).

RELNOTES: `cc_library` targets must have `swift_module` or
`swift_module=ModuleName` in their `tags` in order to be imported into
Swift.
PiperOrigin-RevId: 295823499