Refactor the file containing `swift_common`.
Some of the APIs exported by `swift_common` were only defined in `api.bzl`, which could lead to load cycles if we wanted to export something (such as an aspect) in `swift_common` that also used `swift_common` in its implementation. Out of necessity, many of these APIs had already begun to be moved out of `api.bzl`, so this change completes that process.

`api.bzl` has now been renamed `swift_common.bzl` (which does not matter to clients, since it is an internal implementation file) and *only* contains the `struct` that exports the public APIs.

RELNOTES: None.
PiperOrigin-RevId: 297207799