Rewrite the `swiftc` params file if an argument contains a space.
Bazel does not quote arguments when it writes a multiline params file, but LLVM's command line parser does not distinguish between newlines and other kinds of whitespace, and we don't want to use the shell-quoted params file format from Bazel because it would be more difficult for us to parse. This change ensures that if a line in the incoming params file contains a space, it causes a rewrite which triggers the defensive "quote everything" logic.

RELNOTES: None.
PiperOrigin-RevId: 307876851