Fix up how some direct dependency headers are handled for PCMs.
When compiling a PCM, if the local headers include headers from a direct dependency
that also need additional include paths from their `cc_library` to be found,
ClangImporter wouldn't find them since we were only propagating the headers but not
the rest of the compilation context. Instead of propagating the headers separately,
make sure that all of the direct deps compilation contexts are used when compiling
a PCM, but that only the local context is propagated further.

RELNOTES: None.
PiperOrigin-RevId: 308836734