Allow cacheable swiftmodules option with opt
We've seen cases with envoy-mobile where, when producing a static
framework artifact, neither serialize debugging options flag is passed,
but `-fmodule-map-file` is embedded in the swiftmodule file from the C++
libraries it depends on. This appears to be because the default in the
Swift compiler prefers embedding them in some cases.