Split out implicit module usage from the `swift.use_c_modules` feature.
Create a new `swift.implicit_modules` feature that enables ClangImporter to do an implicit search for modules and use the implicit module cache, and enable it by default. This will allow explicit modules to be used for targets in the workspace while still permitting implicit modules to be used for things like SDK frameworks.

This change also removes the `-fno-implicit-modules` flag, which is only correct for builds where modules from *all* dependencies are able to be compiled explicitly. If some dependency added `-swift.emit_c_module` to its features, this would fail, so it's better to just never pass the flag.

RELNOTES: None.
PiperOrigin-RevId: 310155934