Fix passing -num-threads with -wmo
Previously the assumption was that if `-num-threads` is `1`, only a
single object is emitted. It turns out in swiftc there is a difference
between 0 and 1 to this value, so even with 1 multiple objects are
emitted.

Without this change linking fails when passing this flag like:

```
bazel build examples/xplatform/xctest --swiftcopt=-num-threads --swiftcopt=1 --swiftcopt=-wmo
```