Allow `swift_proto_library` to compile with library evolution enabled.
This feature is not fully supported because the SwiftProtobuf library has not yet been designed to fully support library evolution. The intent of this CL is to allow users building distributable frameworks to use Swift protos as an _implementation-only_ detail of their framework, where those protos would not be exposed to clients in the API. Rely on it at your own risk.

RELNOTES: None.
PiperOrigin-RevId: 314986144