Add a feature to allow Swift targets to skip embedding the `.swiftmodule` file in debug info.
* To disable `.swiftmodule` embedding for a single `swift_{binary,library,test}` target, add `features = ["swift.no_embed_debug_module"]` to that target.
* To disable `.swiftmodule` embedding for all Swift targets in a build package, add `features = ["swift.no_embed_debug_module"]` to the `package(...)` declaration in that BUILD file.
* To disable `.swiftmodule` embedding for an entire build invocation, pass `--features=swift.no_embed_debug_module` on the command line.

(The feature only affects the target it is applied to, not its `deps`. To turn it off universally, use the command line option.)

This is intended to be a workaround for apps with large Swift dependency graphs until Xcode 12 releases a fixed version of `ld64` that supports param files. Using this feature may cause a regression in the debugging experience (e.g., breakpoints, expression evaluation) for modules where it is enabled.

RELNOTES: None.
PiperOrigin-RevId: 322195872