Add `swift.use_tmpdir_for_module_cache` feature
This adds a new feature, that can be enabled by passing
`--features=swift.use_tmpdir_for_module_cache` to the build
command.

This feature, when enabled, will makes the Swift compilation actions use
the shared Clang module cache path written to
`/private/tmp/__build_bazel_rules_swift`. This makes the embedded Clang
module breadcrumbs deterministic between Bazel instances, because they
are always embedded as absolute paths. Note that the use of this cache
is non-hermetic--the cached modules are not wiped between builds, and
won't be cleaned when invoking `bazel clean`; the user is responsible
for manually cleaning them.

Additionally, this can be used as a workaround for a bug in the Swift
compiler that causes the module breadcrumbs to be embedded even though the
`-no-clang-module-breadcrumbs` flag is passed
(https://bugs.swift.org/browse/SR-13275).