Add a feature flag to enable generation from raw proto sources.
The DescriptorSets ProtoInfo exposes don't have source info, so the generated
source don't have the comment. https://github.com/bazelbuild/bazel/issues/9337
is open to attempt to get that, but this provides a way to opt into forcing it.

This does come with a minor risk for cross repository and/or generated proto
files where the protoc command line might not be crafted correctly, so it
remains opt in.

RELNOTES: None
PiperOrigin-RevId: 330538313