Honor the feature flag to enable generation from raw proto sources.
The DescriptorSets ProtoInfo exposes don't have source info, so the generated
source don't have the comment. https://github.com/bazelbuild/bazel/issues/9337
is open to attempt to get that, but this provides a way to opt into forcing it.

This does come with a minor risk for cross repository and/or generated proto
files where the protoc command line might not be crafted correctly, so it
remains opt in.

This is followup to the previous change that did this for swift_proto_library.

NOTE: The current version of swift grpc depended on doesn't include comments,
but the NIO version appears to support it for methods, so landing the work
in advance of things updating to that version.

RELNOTES: None
PiperOrigin-RevId: 330729710