Add `swift.global_module_cache_uses_tmpdir` feature
This adds a new feature, that can be enabled by passing --features=swift.use_global_module_cache and --features=swift.global_module_cache_uses_tmpdir to the build command.

This feature, when enabled, will makes the Swift compilation actions use the shared Clang module cache path written to `/private/tmp/__build_bazel_rules_swift/BIN_DIR/_swift_module_cache`. This makes the embedded Clang module breadcrumbs deterministic between Bazel instances, because they are always embedded as absolute paths. Note that the use of this cache is non-hermetic--the cached modules are not wiped between builds, and won't be cleaned when invoking bazel clean; the user is responsible for manually cleaning them.

Since the source path of modulemaps might be different for the same module, (i.e. multiple checkouts of the same repository, or remote execution) `-fdisable-module-hash` is passed as well. If multiple builds happen at the same time, and different pcm's are generated, this can result in inconsistent builds.

Additionally, this can be used as a workaround for a bug in the Swift compiler that causes the module breadcrumbs to be embedded even though the `-no-clang-module-breadcrumbs` flag is passed (https://bugs.swift.org/browse/SR-13275).