Pass `-no-clang-module-breadcrumbs` to frontend jobs by default
This adds a feature `swift.embed_clang_module_breadcrumbs`, if enabled,
embed absolute paths to the dependent pcm files in the debug info. This
is used as a fallback path for LLDB to reconstruct Clang types from
DWARF, when it wasn't able to import a dependent Clang module from
source. This feature is disabled by default (by passing `-Xfrontend
-no-clang-module-breadcrumbs` to swiftc) to make the compilation
outputs more cacheable.