Delete support for compiler performance stats collection.
This wasn't used frequently enough to rationalize being so deeply woven into the build rules. A better approach might be to use an aspect that runs the same actions in parallel but with the necessary stats flags and collects the data files; while this would double the actions executed, it's better code-health wise. But I don't have any plans to do that in the near-term.

PiperOrigin-RevId: 363887607
(cherry picked from commit 4234e38ce3b4efc0fb0319c96f9052f07bed6775)