Add returns (#634)
Missed with 84286ada38264db60fdfc907bb72863471de6fd7.

Also reverts the hard error on not creating directories. It could try to create the same directory multiple times, and the previous behavior silently errored on those. This can get fixed up in a later change to only try to create each directory once, and to only error if the directory doesn't already exist.