Fix race condition in MakeDirs
If there are 2 concurrent processes creating the same directories, then it will fail

I noticed this on the CI errors of #567, and a co-worker was hitting it
today. I suspect invalidating the worker or other updates triggered big
rebuilds.

I don't know of a perfect repro but, at a 10% rate: invoking Bazel to
build 2 swift libraries in the same BUILD file.