Explicitly pass `alwayslink` when we synthesize the `Objc` provider.
Bazel's implementation of `create_linking_context_from_compilation_outputs` only sets `alwayslink` on a **copy** of the `LibraryToLink` that it merges into the returned `CcLinkingContext`, but **not** on the `LibraryToLink` that is returned in the `CcLinkingOutputs`. This means we can't read it back from that value, as it will always return false.

PiperOrigin-RevId: 380240044
(cherry picked from commit dad2488b29b87fef53e583f49ae019992832ae5a)