Fetch the Swift toolchain as part of the CI job
I'm trying to remove parts from our Docker containers that are only used by one or few projects on CI.

Instead, we can download & extract them during the shell_commands step of the projects. This also allows the projects themselves to control the exact versioning of their dependencies and makes it easier to test upgrades.