Have `swift_clang_module_aspect` always generate the module map for Objective-C targets, instead of using the one generated by Bazel.
The one exception is when a custom module map is provided using the `module_map` attribute; in that case, that module map is still used. But in all cases, the `direct_module_maps` field of `ObjcProvider` is no longer read by the Swift build rules.

This change includes some very unfortunate workarounds to address J2ObjC's tightly coupled relationship with the Obj-C logic in Bazel; even though the module map is never used, Bazel must still generate it since umbrella header generation is coupled to it. In doing so, however, this change also updates module map generation to support expanding directory artifacts and listing umbrella headers (which we may want to support in `swift_interop_hint` separately in the future).

PiperOrigin-RevId: 395704174