Fix incremental compilation with derived files
For incremental compilation we pass an output file map to compile actions. Since we have 2 actions with the derived files feature, they were fighting to write the same set of files. In this case we now pass a separate output file map to the derived files action that only contains the swiftmodules, and one that only contains the objects to the standard compile.

In addition, the derived files action uses a separate staging directory, as the various `swiftdeps` and `priors` files conflict.

Finally, we only set `incremental_inputs` for the swift module producing action.