Add a flag to allow tools to be compiled as universal binaries
This adds a new boolean flag
`--@build_bazel_rules_swift//swift:universal_tools` that when used will
force all the tools used in this repository to be compiled into
universal binaries. This allows execution running on different macOS
platforms (`arm64` and `x86_64`) to be compatible with each other and
can run natively on both.