Ensure quote_includes are set for swift_library's CcInfo
This mimics the behavior of cc_library.

When consuming CcInfo from a swift_library, if the library is in an external
repo, the quote_includes paths are not set.
In the case of a cc_library, both the workspace root and the bindir +
workspace_root are set.

If the workspace_root isn't set (in case of a non external repo), only
include the bin_dir.

Signed-off-by: Steeve Morin <steeve@zen.ly>