Reland: Retrieve single arch cpu from `cc_toolchain` instead of from `ctx.fragments.apple`
`ctx.fragments.apple.single_arch_cpu` is returning the default macos
cpu for the host platform for tools while it should return the cpu of
the execution platform.

There is a caveat with this workaround: When building a `swift_library`
directly without a `--cpu` flag, the `cpu` value would be the cpu value
of the exec platform, regardless of the `--apple_platform_type` value.