Patch rules_proto to make `protoc` universal everywhere
The `protoc` executable is also used by native rules (e.g.
`proto_library`), and is referred to as `@com_google_protobuf//:protoc`.
The `protoc_wrapper` we previously used only worked for rules in
rules_swift but not elsewhere - the universal target has to have the
label `@com_google_protobuf//:protoc`.